TW-331: Notifications. Refactor
