+++ src/lib/notifications/index.ts
@@ -1,7 +1,7 @@
-export { Notifications } from './components/notifications/notifications';
-export { NotificationsBell } from './components/notifications-bell/notifications-bell';
-export { NotificationsItem } from './components/notifications-item/notifications-item';
-export { NotificationsSettings } from './components/notifications-settings/notifications-settings';
+export { Notifications } from './components/notifications';
+export { NotificationsBell } from './components/bell';
+export { NotificationsItem } from './components/item';
+export { NotificationsSettings } from './components/settings';
 
 export { loadNotificationsAction } from './store/actions';
 export { notificationsEpics } from './store/epics';
